requirements.txt
runpod>=1.2.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.50.0
Pillow>=9.5.0
requests>=2.28.0
accelerate>=0.20.0
